requirements.txt
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
cryptography==36.0.1
dnspython==2.2.0
Flask==2.0.2
Flask-PyMongo==2.3.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
nose==1.3.7
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pymongo==4.0.1
python-dotenv==0.19.2
Random-Word==1.0.7
requests==2.27.1
toml==0.10.2
urllib3==1.26.8
Werkzeug==2.0.3
